feat(convention): create a convention of `container` (#115)
The Open Container Initiative (OCI) is an open governance structure for the express purpose of creating open
industry standards around container formats and runtimes.

The Compose Specification is a developer-focused standard for defining cloud and platform agnostic container-
based applications. The compose specification establishes a standard for the definition of multi-container
platform-agnostic applications.